[BIT-010] User Project Assignment

+++ bitfrost/src/main/java/com/janne/bitfrost/controllers/ProjectController.java
@@ -2,20 +2,23 @@ package com.janne.bitfrost.controllers;
 
 import com.janne.bitfrost.entities.Project;
 import com.janne.bitfrost.entities.Topic;
+import com.janne.bitfrost.services.AuthService;
 import com.janne.bitfrost.services.ProjectService;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.ArrayList;
 import java.util.List;
 
+@Slf4j
 @RestController
 @RequestMapping("/projects")
 @RequiredArgsConstructor
 public class ProjectController {
 
     private final ProjectService projectService;
+    private final AuthService authService;
 
     @GetMapping
     public ResponseEntity<List<Project>> getProjects() {
@@ -24,17 +27,40 @@ public class ProjectController {
 
     @PostMapping
     public ResponseEntity<Project> createProject(@RequestBody Project project) {
-        return ResponseEntity.ok(projectService.createProject(project));
+        Project buildProject = projectService.createProject(project);
+        projectService.assignUserToProject(authService.getUser().getUuid(), buildProject.getProjectTag());
+        return ResponseEntity.ok(buildProject);
     }
 
-    @PostMapping("/{projectId}/topic")
-    public ResponseEntity<Project> createTopic(@PathVariable String projectId, @RequestBody Topic topic) {
-        return ResponseEntity.ok(projectService.addTopic(projectId, topic));
+    @PostMapping("/{projectTag}/topic")
+    public ResponseEntity<Project> createTopic(@PathVariable String projectTag, @RequestBody Topic topic) {
+        authService.assertUserHasProjectAccess(projectTag);
+        return ResponseEntity.ok(projectService.addTopic(projectTag, topic));
+    }
+
+    @GetMapping("/owned")
+    public ResponseEntity<List<Project>> getOwnedProjects() {
+        return ResponseEntity.ok(authService.getUser().getAssignedProjects().stream().toList());
     }
 
     @DeleteMapping("/{projectTag}")
     public ResponseEntity<Void> deleteProject(@PathVariable String projectTag) {
+        authService.assertUserHasProjectAccess(projectTag);
         projectService.deleteProject(projectTag);
+        return ResponseEntity.ok().build();
+    }
+
+    @PostMapping("/allow/{projectTag}")
+    public ResponseEntity<Void> allowAccess(@PathVariable String projectTag, @RequestParam String userId) {
+        authService.assertUserHasProjectAccess(projectTag);
+        projectService.assignUserToProject(userId, projectTag);
+        return ResponseEntity.ok().build();
+    }
+
+    @DeleteMapping("/revoke/{projectTag}")
+    public ResponseEntity<Void> revokeAccess(@PathVariable String projectTag, @RequestParam String userId) {
+        authService.assertUserHasProjectAccess(projectTag);
+        projectService.revokeUserFromProject(userId, projectTag);
         return ResponseEntity.ok().build();
     }
 }
